Create only one schema definition
Instead of carrying a schema per table I believe it is easier to use
only one sql file for various reasons.

It is exactly the same as what mysqldump would use:
 * mysqldump -u root --no-data ouija

It removes the wondering if there's a reason for them to be separate
(e.g. some apply to ouija while others to seta).