Add information of new jobs + count new_jobs properly
We want to see in the output what new jobs are being added.

Moving new_jobs outside of the finally statement fixes the issue were
failed changes would also be counted as new jobs since the finally
statement gets executed for both the good case as well as when there is
an exception.